Move WebRTC.isSupported property to mediaHandlerFactory method.
Instead of checking SIP.WebRTC.isSupported, SIP.js checks
mediaHandlerFactory.isSupported(). If the method is not defined, it is
considered to have returned true, making it optional for a custom
mediaHandlerFactory.

This has the side effect of binding SIP.WebRTC constructor methods only after
SIP.WebRTC.isSupported is called.

fixes #37