Fixes the way utf8 string length is counted.
When sending SIP messages with "special characters" e.g. "äöü",
the "Content-Length" of the body needs to be correct. (Content-length
mis-match). Before this change, the byte count was wrong e.g.

"ä" => 1 instead of 2

I wrote a couple of examples in "SpecSIPMessage" and double checked them
with MRI (Ruby 2.0.0).

Thanks to [Tgr](http://stackoverflow.com/a/2848483/980524)