Complete the SIP-Message specs.
I noticed some specs in SpecSIPMessage were not yet implemented.
They assert what was written in the corresponding 'it' blocks, with one
exception:

if a header is neither in headers nor extraHeaders,

    .getHeaders()

returns an empty array, but the spec said it should return 'undefined'.
I think an emtpy array makes more sense though. Let me know if you
disagree and I'll change it.