Grammar: fix quoted_string_clean rule
In theory, the rule returns what's between the quotes in the input.

In practice, it was slicing off the first and last characters of the input,
then returning the result. This only worked if there wasn't any SWS before
the quoted string.

Now, it actually returns what's between the quotes.