Fix FireFox turn url initialization NS_ERROR_UNEXPECTED
For a turn configuration like:

    turnServers: [
      {
        urls: ['turn:stun.example.com:3478', 'turn:stun.example.com:3478?transport=tcp', 'turns:stun.example.com:443?transport=tcp'],
        username: 'alice',
        password: 'root'
      }
    ],

the `servers` variable in:

    this.peerConnection = new SIP.WebRTC.RTCPeerConnection({'iceServers': servers}, this.RTCConstraints);

    [
     {
        credential: "root",
        url: [
            "turn:stun.example.com:3478",
            "turn:stun.example.com:3478?transport=tcp",
            "turns:stun.example.com:443?transport=tcp"
        ],
        username: "alice"
      }
    ]

I couldn't any documentation or spec which specifies if arrays are "allowed"
for the `url` property but it seems like Chrome swallows it anyway.
However FireFox (34) spits out `NS_ERROR_UNEXPECTED`.

With the proposed change `servers` would contain an object for each url
(with credential and username). This works in FF and Chrome.