Don't mutate user submitted options object in UA.invite.
I ran into the following situation today:
After a call to UA.invite:

    userAgent.invite("xyz@test.org", myopts);

The myopts object had a `extraHeaders` properties added.
(Session L1600).

The problem is if you have a setup like this:

      var myopts = {
        media: {
          constraints: {
            audio: true,
            video: true
          },
          render: {
            remote: {
              video: document.getElementById("remoteVideo")
            },
            local: {
              video: document.getElementById("localVideo")
            }
          }
        }
      };

    document.getElementById("callButton").onclick = function() {
      session = userAgent.invite("xyz@test.org", myopts);
      console.log(myopts);
    };

where you re-use the myopts object, you keep `push`ing extraHeaders
`Contact:` and `Allow:` headers and finally sending these with your invites.
I thought that a copy of the user-submitted object might be the cleanest
solution to this problem. But of course I'm open to other suggestions (and
alternative implementations of the copy functionality).