Update Client (and to some extent Server) Context to be friendlier with options.
I'm not entirely satisfied with what I did to ServerContext.
I removed the unique helper function, but ended up making
everything a little bigger in the process.  Plus, the actual
`reply` method doesn't use the same pattern as the others
when it comes to options.

Some of that will probably be fixed with an Options module
to dedupe the handling across SIP.js