Remove SDP mangling from sendReinvite; move hold functionality to MediaHandler
local_hold and remote_hold on the MediaHandler are similar to local_hold and
remote_hold on the Session. If the remote_hold value differs between the
MediaHandler and the Session, then the Session should update its remote_hold
property and emit an event. The local_hold property on the MediaHandler is
used to selectively disable media sections in the SDP.