Bind holdModifier to the SessionDescriptionHandler instance
When putting a call on hold, sessionDescriptionHandler.holdModifier is
added to the modifiers. The modifier is an instance method, and as such
we need to bind it to the instance.

It so happens the default WebRTC holdModifier does not reference 'this',
so things work as expected. However for a custom session description
handler this may not hold true.