Use babel-loader to transpile code
Switch from eslint-loader to babel-loader to allow using modern constructs
such as arrow functions or ES6 classes while still retaining "compatibility"
with older browsers. It's worth mentioning browsers which support WebRTC
already ship with ES6 support, but at least we won't trigger parser errors
on the older browsers.

The cost in terms of runtime size is negligeable, the minified code grows
by less than 400 bytes as a result of the switch. Note: we don't pull in
any polyfills with this commit, so only ES6 constructs are supported, not
using new classes such as Map or Set.

We still perform linting as part of the build, by invoking eslint in the
prebuild step.