Add P-Asserted-Identity header support
RFC 3325 indicates:
Typically, a user agent renders the value of a P-Asserted-Identity
header field that it receives to its user.  It may consider the identity
provided by a Trust Domain to be privileged, or intrinsically more
trustworthy than the From header field of a request.  However, any
specific behavior is specific to implementations or services.

The header has been added so that UA's can use it instead of the
remoteIdentity if this behavior is required.