Install jasmine-core from package.json, not just travis
If we list jasmine-core (which we need to run tests) in devDependencies
there is no need for a travis-specific hook. This also ensures we don't
have a version mismatch when we decide to upgrade jasmine.

Also on trusty, the default gcc version is already 4.8 so save time by
not downloading additional packages via apt.