Transport: Allow infinite reconnection
The current behavior is to try each server a given number of times
depending on the maxReconnectionAttempts, then fail permanently.

We could have network situations where the softphone is disconnected
from the Internet. Users expect the softphone to come back online as
soon as possible without the need to reload. That is also a standard
bevhaviour among hardware IP Phones and other softphones.

This is achieved thanks to this patch and the introduction of the
alwaysReconnect parameter.