Refresh of Transactions
  - Fixes #517
  - Working 2xx ACK retransimissions
  - Proper outgoing CANCEL in transactions
  - Correct and complete timers (for reliable transports)
  - Abstract transaction user (remove RequestSender dependency)
  - Update and integrate TransportError
  - Move message creation to sip-message
  - Fixup some types in sip-message
  - A bunch of minor bug fixes
  - Pedantic state machines
  - And more...

Aside from a little bit of polishing with an eye toward code maintenance,
the goal was to move the Transaction interface to a place where some
other outstanding issues can be addressed without resorting to hacks.
In particular, looking to build on this work to deal with some known
race condition issues in Session that are hard to solve with the current
Transaction interface (and implementation).