Integrate Contexts and Core code with existing library
The new Contexts and Core code is "hidden" behind a
new UA option, "experimentalFeatures", which is off (false)
by default.  If "experimentalFeatures" is enabled (true),
the new Contexts and Core code is utilized.

The intention is that the new code is backward compatible
with the existing library and one should be able to enable it
without needing to make any changes, but the difference
in the code that runs is non-trivial and not yet well tested.