Subscription auto refresh fix
This existing implementation resets the refresh timeout
to be 90% of whatever the expire on the latest NOTIFY is.
Thus on every NOTIFY received the refresh will occur closer
to the expiration time of the subscription. For subscriptions
with frequent NOTIFYs, the refresh eventually occurs after
the expires.

The new implementation refreshes at 90% of the original expiration.