InviteServerContext - allow sending of 183 w/SDP
If progress called with 183 and no body, get an offer/answer
from the SDH and send the SDP in the body of the 183.
It is the de facto industry standard to utilize 183 with SDP
to provide "early media", so this seems like a reasonable
behavior.

Added this so could write tests for the bug fixed in prior commit.
Also helpful moving to address "early media" in InviteClientContext.