New API - deprecate autoStart
The UserAgentOption autoStart currently
defaults to `true`. While autoStart was
apparently an attempt to simplify library
usage, it creates an issue where it is
difficult to know when the transport
is connected after starting the UserAgent
so ends up require more complicated
code to be written at the end of the day
In effect, it is self defeating.

It now defaults to `false` and has been
marked at deprecated.

This commit addresses:
#738
#739