New API - preloaded route set option change
Replaced "usePreloadedRoute" UserAgentOption with "preloadedRouteSet".
- Preloaded route now set in one place
- Removed dependency on Transport
- Allows for multiple Route headers

So going forward, if you want to use a preload route set with the new API
you should specify it using this user agent option. It the past the you
would need to set "usePreloadedRoute" to "true" and then set the
"server.sipUri" on your Transport to insert a preloaded route.